Split practice-controller.test.ts by action group
